[proliferate] extract session directory helpers
