+++ desktop/src/hooks/app/lifecycle/use-export-running-agent-count.ts
@@ -1,10 +1,8 @@
 import { useEffect } from "react";
 import { isSessionSlotBusy } from "@/lib/domain/sessions/activity";
+import { activitySnapshotFromDirectoryEntry } from "@/lib/domain/sessions/directory/directory-activity";
 import { useTauriWindowActions } from "@/hooks/access/tauri/use-window-actions";
-import {
-  activitySnapshotFromDirectoryEntry,
-  useSessionDirectoryStore,
-} from "@/stores/sessions/session-directory-store";
+import { useSessionDirectoryStore } from "@/stores/sessions/session-directory-store";
 
 type SessionEntries = ReturnType<typeof useSessionDirectoryStore.getState>["entriesById"];
