+++ desktop/src/hooks/sessions/lifecycle/use-session-history-hydration.ts
@@ -29,11 +29,9 @@ import { uniqueMeasurementOperationIds } from "@/lib/infra/measurement/operation
 import { batchSessionStoreWrites } from "@/lib/infra/scheduling/react-batching";
 import { scheduleAfterNextPaint } from "@/lib/infra/scheduling/schedule-after-next-paint";
 import { fetchSessionHistory } from "@/lib/workflows/sessions/session-runtime";
+import { activityFromTranscript } from "@/lib/domain/sessions/directory/directory-activity";
 import { useLinkedSessionMounting } from "@/hooks/chat/subagents/use-linked-session-mounting";
-import {
-  activityFromTranscript,
-  useSessionDirectoryStore,
-} from "@/stores/sessions/session-directory-store";
+import { useSessionDirectoryStore } from "@/stores/sessions/session-directory-store";
 import { getSessionRecord } from "@/stores/sessions/session-records";
 import { useSessionTranscriptStore } from "@/stores/sessions/session-transcript-store";
 import type { SessionRuntimeRecord } from "@/stores/sessions/session-types";
