+++ desktop/src/hooks/sessions/session-runtime-helpers.ts
@@ -1,11 +1,9 @@
 import type { SessionStreamHandle } from "@anyharness/sdk";
 import { resolveSessionViewState } from "@/lib/domain/sessions/activity";
+import { activitySnapshotFromDirectoryEntry } from "@/lib/domain/sessions/directory/directory-activity";
 import { isPendingSessionId } from "@/lib/workflows/sessions/session-runtime";
 import { isCurrentSessionStreamHandle } from "@/lib/access/anyharness/session-stream-handles";
-import {
-  activitySnapshotFromDirectoryEntry,
-  useSessionDirectoryStore,
-} from "@/stores/sessions/session-directory-store";
+import { useSessionDirectoryStore } from "@/stores/sessions/session-directory-store";
 
 export function shouldReconnectStream(sessionId: string): boolean {
   const entry = useSessionDirectoryStore.getState().entriesById[sessionId];
