+++ desktop/src/hooks/sessions/workflows/use-session-summary-actions.ts
@@ -15,14 +15,12 @@ import {
   type PendingSessionConfigChange,
 } from "@/lib/domain/sessions/pending-config";
 import { buildSessionSlotPatchFromSummary } from "@/lib/domain/sessions/summary";
+import { activityFromTranscript } from "@/lib/domain/sessions/directory/directory-activity";
 import { batchSessionStoreWrites } from "@/lib/infra/scheduling/react-batching";
 import { persistDefaultSessionModePreference } from "@/hooks/sessions/session-mode-preferences";
 import { clearPendingConfigRollbackCheck } from "@/hooks/sessions/session-runtime-pending-config";
 import { useWorkspaceSurfaceLookup } from "@/hooks/workspaces/use-workspace-surface-lookup";
-import {
-  activityFromTranscript,
-  useSessionDirectoryStore,
-} from "@/stores/sessions/session-directory-store";
+import { useSessionDirectoryStore } from "@/stores/sessions/session-directory-store";
 import { getSessionRecord } from "@/stores/sessions/session-records";
 import { useSessionTranscriptStore } from "@/stores/sessions/session-transcript-store";
 import { trackWorkspaceInteraction } from "@/stores/preferences/workspace-ui-store";
