+++ desktop/src/hooks/workspaces/tabs/workspace-header-live-slots-selector.ts
@@ -1,4 +1,4 @@
-import type { SessionDirectoryEntry } from "@/stores/sessions/session-types";
+import type { SessionDirectoryEntry } from "@/lib/domain/sessions/directory/directory-entry";
 
 type SessionDirectoryStoreSnapshot = {
   entriesById: Record<string, SessionDirectoryEntry>;
