+++ desktop/src/stores/sessions/session-directory-store.test.ts
@@ -1,9 +1,7 @@
 import { createTranscriptState, type PendingInteraction } from "@anyharness/sdk";
 import { beforeEach, describe, expect, it, vi } from "vitest";
-import {
-  activitySnapshotFromDirectoryEntry,
-  useSessionDirectoryStore,
-} from "@/stores/sessions/session-directory-store";
+import { activitySnapshotFromDirectoryEntry } from "@/lib/domain/sessions/directory/directory-activity";
+import { useSessionDirectoryStore } from "@/stores/sessions/session-directory-store";
 
 describe("session directory store invariants", () => {
   beforeEach(() => {
